fix(command-board): move non-async exports out of server action files
- retry-utils.ts moved to lib/ (was in actions/, causing Next.js server action validation error)
- CONNECTION_STYLES/CONNECTION_COLORS moved to lib/connection-constants.ts
- Updated imports in annotations.ts, boards.ts, update-entity.ts, connection-dialog.tsx
