+++ apps/app/app/(authenticated)/command-board/actions/annotations.ts
@@ -4,7 +4,7 @@ import { database } from "@repo/database";
 import { revalidatePath } from "next/cache";
 import { requireTenantId } from "../../../lib/tenant";
 import type { BoardAnnotation } from "../types/board";
-import { withRetry } from "./retry-utils";
+import { withRetry } from "../lib/retry-utils";
 
 // ============================================================================
 // Types
@@ -231,20 +231,5 @@ export async function deleteAnnotation(
 // Connection Type Helpers
 // ============================================================================
 
-/** Available connection styles for manual connections */
-export const CONNECTION_STYLES = [
-  { value: "solid", label: "Solid", description: "Standard solid line" },
-  { value: "dashed", label: "Dashed", description: "Dashed line pattern" },
-  { value: "dotted", label: "Dotted", description: "Dotted line pattern" },
-] as const;
-
-/** Default colors for manual connections */
-export const CONNECTION_COLORS = [
-  { value: "#9ca3af", label: "Gray", description: "Neutral gray" },
-  { value: "#3b82f6", label: "Blue", description: "Primary blue" },
-  { value: "#22c55e", label: "Green", description: "Success green" },
-  { value: "#f59e0b", label: "Amber", description: "Warning amber" },
-  { value: "#ef4444", label: "Red", description: "Error red" },
-  { value: "#8b5cf6", label: "Purple", description: "Accent purple" },
-  { value: "#ec4899", label: "Pink", description: "Accent pink" },
-] as const;
+// CONNECTION_STYLES and CONNECTION_COLORS moved to ../lib/connection-constants.ts
+// (Server Action files cannot export non-async-function values)
